+++ script/database/compare_database_application.py
@@ -3,10 +3,6 @@ import argparse
 import csv
 import logging
 import os
-import sys
-
-new_path = os.path.normpath(os.path.join(os.path.dirname(__file__), ".."))
-sys.path.append(new_path)
 
 logging.basicConfig(level=os.environ.get("LOGLEVEL", "INFO"))