+++ script/database/db_restore.py
@@ -7,9 +7,6 @@ import os
 import sys
 from subprocess import check_output
 
-new_path = os.path.normpath(os.path.join(os.path.dirname(__file__), ".."))
-sys.path.append(new_path)
-
 logging.basicConfig(level=os.environ.get("LOGLEVEL", "INFO"))
 
 _logger = logging.getLogger(__name__)