+++ script/docker/docker_update_version.py
@@ -11,7 +11,7 @@ new_path = os.path.normpath(
 )
 sys.path.append(new_path)
 
-from script.git_tool import GitTool
+from script.git.git_tool import GitTool
 
 _logger = logging.getLogger(__name__)