+++ script/git/git_merge_repo_manifest.py
@@ -5,10 +5,12 @@ import logging
 import os
 import sys
 
-new_path = os.path.normpath(os.path.join(os.path.dirname(__file__), ".."))
+new_path = os.path.normpath(
+    os.path.join(os.path.dirname(__file__), "..", "..")
+)
 sys.path.append(new_path)
 
-from script.git_tool import GitTool
+from script.git.git_tool import GitTool
 
 _logger = logging.getLogger(__name__)