+++ script/git/git_repo_update_group.py
@@ -7,10 +7,12 @@ import sys
 import git
 from git import Repo
 
-new_path = os.path.normpath(os.path.join(os.path.dirname(__file__), ".."))
+new_path = os.path.normpath(
+    os.path.join(os.path.dirname(__file__), "..", "..")
+)
 sys.path.append(new_path)
 
-from script.git_tool import GitTool
+from script.git.git_tool import GitTool
 
 _logger = logging.getLogger(__name__)