+++ script/git/repo_remove_auto_install.py
@@ -8,10 +8,12 @@ from pathlib import Path
 from git import Repo  # pip install gitpython
 from git.exc import GitCommandError
 
-new_path = os.path.normpath(os.path.join(os.path.dirname(__file__), ".."))
+new_path = os.path.normpath(
+    os.path.join(os.path.dirname(__file__), "..", "..")
+)
 sys.path.append(new_path)
 
-from script.git_tool import GitTool
+from script.git.git_tool import GitTool
 
 _logger = logging.getLogger(__name__)