+++ script/poetry/poetry_add_build_dependency.sh
@@ -4,7 +4,7 @@
 # poetry export -f ./.venv/build_dependency.txt --dev | poetry run -- pip install -r /dev/stdin
 retVal=$?
 if [[ $retVal -ne 0 ]]; then
-    echo "Error ./script/poetry_add_build_dependency.sh"
+    echo "Error ./script/poetry/poetry_add_build_dependency.sh"
     cat ./.venv/build_dependency.txt
     exit 1
 fi